Address PR #877 review comments
- ExportMenuItem: call hideMenu in a `finally` so the export menu always closes
  even if the onClick handler rejects/throws (Copilot).
- Extract the duplicated NoopTransition into a shared
  components/NoopTransition.tsx, used by IntegrationCard and ImageEditingDialog
  (Greptile, NoopTransition duplicated across files).

Not changed (with rationale):
- MapViewer Rectangle/Polygon casts (Greptile P1): tsc is green without them.
  Only MapContainer/TileLayer/Circle produce React 19 ref-generic errors and
  thus need the cast; casting Rectangle/Polygon would add no-op `as unknown as`
  casts that suppress real type-checking.
- forwardRef -> ref-as-prop migration (Greptile P2): forwardRef still works in
  React 19 (no active runtime warning), the React Compiler is deferred in this
  PR, and migrating MUI transition slots is behaviour-sensitive (ref forwarding
  + focus, not verifiable in jsdom). Deferred to the compiler-enablement work;
  the now-shared NoopTransition documents this.

tsc, lint, prod build, and the full jsdom suite (371 files, 2018 tests) pass.
